+++ tests/Tsufeki/Tenkawa/FunctionalTest.php
@@ -14,6 +14,7 @@ use Tsufeki\Tenkawa\Server\Logger\LevelFilteringLogger;
 use Tsufeki\Tenkawa\Server\Logger\StreamLogger;
 use Tsufeki\Tenkawa\Server\ServerPlugin;
 use Tsufeki\Tenkawa\Server\Tenkawa;
+use Tsufeki\Tenkawa\WebMozartAssert\WebMozartAssertPlugin;
 
 /**
  * @coversNothing
@@ -50,6 +51,7 @@ class FunctionalTest extends TestCase
                     new PhonyPlugin(),
                     new ProphecyPlugin(),
                     new MockeryPlugin(),
+                    new WebMozartAssertPlugin(),
                 ]
             );